Change type of SIZELIMIT to size_t
Newer gcc are picky about signed aund unsigned pointer types for a good
reason. Change the type of SIZELIMIT to size_t as preperatory work for
a huge signed/unsigned pointer cleanup.

+++ Src/config.h
@@ -347,7 +347,7 @@ extern int MIDDLEMAN;
 extern int AUTOBLOCK;
 extern int STATSDETAILS;
 extern char FORWARDTO[];
-extern int SIZELIMIT;
+extern size_t SIZELIMIT;
 extern int INFLATEMAX;
 extern int MAXRANDHOPS;
 extern int BINFILTER;
